requirements.txt
endee>=0.1.10
sentence-transformers>=2.7.0
python-dotenv>=1.0.1
click>=8.0.0
requests>=2.31.0
pydantic==1.10.26
llama-index-core>=0.9.0
openai>=1.3.0
anthropic>=0.7.0
PyPDF2>=3.0.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
